deinline nmod_poly IO
